requirements.txt
langchain
langgraph
langchain-google-genai
google-api-python-client
google-auth
python-dotenv
pydantic
fastapi
uvicorn
streamlit
